added retry

+++ pkg/generate/cleanup/cleanup.go
@@ -1,6 +1,8 @@
 package cleanup
 
 import (
+	"strconv"
+
 	"github.com/go-logr/logr"
 	kyverno "github.com/kyverno/kyverno/pkg/api/kyverno/v1"
 	dclient "github.com/kyverno/kyverno/pkg/dclient"
@@ -14,13 +16,31 @@ func (c *Controller) processGR(gr kyverno.GenerateRequest) error {
 
 	// 2- The trigger resource is deleted, then delete the generated resources
 	if !ownerResourceExists(logger, c.client, gr) {
-		if err := deleteGeneratedResources(logger, c.client, gr); err != nil {
-			return err
+		deleteGR := false
+		// check retry count in annotaion
+		grAnnotations := gr.Annotations
+		if val, ok := grAnnotations["generate.kyverno.io/retry-count"]; ok {
+			retryCount, err := strconv.ParseUint(val, 10, 32)
+			if err != nil {
+				logger.Error(err, "unable to convert retry-count")
+				return err
+			}
+
+			if retryCount >= 5 {
+				deleteGR = true
+			}
+			// sleepCountInt := int(sleepCountInt64)
 		}
-		// - trigger-resource is deleted
-		// - generated-resources are deleted
-		// - > Now delete the GenerateRequest CR
-		return c.control.Delete(gr.Name)
+
+		if deleteGR {
+			if err := deleteGeneratedResources(logger, c.client, gr); err != nil {
+				return err
+			}
+			// - trigger-resource is deleted
+			// - generated-resources are deleted
+			// - > Now delete the GenerateRequest CR
+			return c.control.Delete(gr.Name)
+		}
 	}
 	return nil
 }
